updated file structure and methods
